Add `EnvoyFilter` for rate-limiting
Add an `EnvoyFilter` custom resource to our deployment file that
allows only a maximum of 10 incoming requests towards our
application.

Specifically, every 20 seconds, 5 more requests will be added to
the "bucket" of allowed requests. The bucket has a maximum capacity
of 10 requests in total.

This is to implement our rate-limiting use-case.